MultiNode Propagation V1
Early Version of propagation for multiple parent and daughter nodes. Currently the Pi Evidence (downward propagation) still needs some work, but this should be fixed in the next update. Once the networks functionality is secured, work will be done on getting the UI working with the network behind the scenes