MultiNode Propagation V2 - Functional
We can now propagate evidence among nodes with multiple parents and daughters. The network can use a lot of optimisations. There are many steps that use excessive computational steps and there are many recomputations of the same thing. I will try to work on that after I work on the interface. Top priority now is getting the UI to be operable. -  Another thing to do would be to create more examples.